refactor: extract temporary fix for easier suppression/removal
